Separated `labelTypes` into separate file
